Fix key remove bug
If you overwrote an existing key which was in the original configuration file and then set it to nil you get nil value when you wanted to tead the key for the first time but after application relaunch you got the original configuration value.

Now you will get the original configuration value for the first time too.